chore(core): refactor events to bring related code together
