+++ src/bp/core/api.ts
@@ -1,4 +1,5 @@
 import * as sdk from 'botpress/sdk'
+import { EventEngine } from 'core/events/event-engine'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { NextFunction, Request, Response } from 'express'
 import { inject, injectable } from 'inversify'
@@ -29,7 +30,6 @@ import { HookService } from './services/hook/hook-service'
 import { JobService } from './services/job-service'
 import { KeyValueStore } from './services/kvs'
 import { MediaServiceProvider } from './services/media'
-import { EventEngine } from './services/middleware/event-engine'
 import { StateManager } from './services/middleware/state-manager'
 import { NotificationsService } from './services/notification/service'
 import RealtimeService from './services/realtime'
