+++ src/bp/core/botpress.ts
@@ -1,5 +1,7 @@
 import * as sdk from 'botpress/sdk'
 import lang from 'common/lang'
+import { addStepToEvent, EventCollector, StepScopes, StepStatus } from 'core/events/event-collector'
+import { EventEngine } from 'core/events/event-engine'
 import { copyDir } from 'core/misc/pkg-fs'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
@@ -40,8 +42,6 @@ import { SessionIdFactory } from './services/dialog/session/id-factory'
 import { HintsService } from './services/hints'
 import { Hooks, HookService } from './services/hook/hook-service'
 import { LogsJanitor } from './services/logs/janitor'
-import { addStepToEvent, EventCollector, StepScopes, StepStatus } from './services/middleware/event-collector'
-import { EventEngine } from './services/middleware/event-engine'
 import { StateManager } from './services/middleware/state-manager'
 import { MigrationService } from './services/migration'
 import { MonitoringService } from './services/monitoring'
