+++ src/bp/core/database/tables/bot-specific/index.ts
@@ -3,6 +3,6 @@ export * from './logs'
 export * from './ghost_content'
 export * from './notifications'
 export * from './kvs'
-export * from './events'
+export * from 'core/events/event-table'
 export * from './tasks'
 export * from './users'
