+++ src/bp/core/events/event-collector.ts
@@ -2,14 +2,13 @@ import * as sdk from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
 import Database from 'core/database'
 import { EventRepository } from 'core/repositories'
+import { SessionIdFactory } from 'core/services/dialog/session/id-factory'
 import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import Knex from 'knex'
 import _ from 'lodash'
 import moment from 'moment'
 import ms from 'ms'
-
-import { SessionIdFactory } from '../dialog/session/id-factory'
 
 type BatchEvent = sdk.IO.StoredEvent & { retry?: number }
