+++ src/bp/core/events/middleware.ts
@@ -1,8 +1,7 @@
 import * as sdk from 'botpress/sdk'
+import { addStepToEvent, StepScopes, StepStatus } from 'core/events/event-collector'
 import _ from 'lodash'
 import ms from 'ms'
-
-import { addStepToEvent, StepScopes, StepStatus } from './event-collector'
 
 interface MiddlewareChainOptions {
   timeoutInMs: number
