+++ src/bp/core/logger/logger.ts
@@ -2,9 +2,9 @@ import { AxiosError } from 'axios'
 import { IO, Logger, LoggerEntry, LoggerLevel, LoggerListener, LogLevel } from 'botpress/sdk'
 import chalk from 'chalk'
 import { Metric } from 'common/monitoring'
+import { addLogToEvent } from 'core/events/event-collector'
 import { IDisposable } from 'core/misc/disposable'
 import { BotService } from 'core/services/bot-service'
-import { addLogToEvent } from 'core/services/middleware/event-collector'
 import { incrementMetric } from 'core/services/monitoring'
 import { InvalidParameterError } from 'errors'
 import { EventEmitter2 } from 'eventemitter2'
