+++ src/bp/core/services/auth/auth-service.ts
@@ -4,6 +4,7 @@ import { AuthStrategy, AuthStrategyBasic } from 'core/config/botpress.config'
 import { ConfigProvider } from 'core/config/config-loader'
 import Database from 'core/database'
 import { StrategyUserTable } from 'core/database/tables/server-wide/strategy_users'
+import { EventEngine } from 'core/events/event-engine'
 import { getMessageSignature } from 'core/misc/security'
 import { ModuleLoader } from 'core/module-loader'
 import { StrategyUsersRepository } from 'core/repositories/strategy_users'
@@ -22,7 +23,6 @@ import { TYPES } from '../../types'
 import { SessionIdFactory } from '../dialog/session/id-factory'
 import { JobService } from '../job-service'
 import { KeyValueStore } from '../kvs'
-import { EventEngine } from '../middleware/event-engine'
 import { WorkspaceService } from '../workspace-service'
 
 import StrategyBasic from './basic'
