+++ src/bp/core/services/converse.ts
@@ -1,5 +1,6 @@
 import { IO } from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
+import { EventEngine } from 'core/events/event-engine'
 import { UserRepository } from 'core/repositories'
 import { TYPES } from 'core/types'
 import { InvalidParameterError } from 'errors'
@@ -10,8 +11,6 @@ import _ from 'lodash'
 import ms from 'ms'
 
 import { Event } from '../sdk/impl'
-
-import { EventEngine } from './middleware/event-engine'
 
 export const converseApiEvents = new EventEmitter2()
