+++ src/bp/core/services/dialog/instruction/strategy.ts
@@ -1,10 +1,10 @@
 import { IO, Logger } from 'botpress/sdk'
 import { parseActionInstruction } from 'common/action'
 import { ActionServer } from 'common/typings'
+import { EventEngine } from 'core/events/event-engine'
 import ActionServersService from 'core/services/action/action-servers-service'
 import ActionService from 'core/services/action/action-service'
 import { CMSService } from 'core/services/cms'
-import { EventEngine } from 'core/services/middleware/event-engine'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
 import { NodeVM } from 'vm2'
