+++ src/bp/core/services/services.inversify.ts
@@ -1,5 +1,6 @@
 import { IO } from 'botpress/sdk'
 import LicensingService from 'common/licensing-service'
+import { EventEngine } from 'core/events/event-engine'
 import { DialogContainerModule } from 'core/services/dialog/dialog.inversify'
 import { CEJobService, JobService } from 'core/services/job-service'
 import { ContainerModule, interfaces } from 'inversify'
@@ -24,7 +25,6 @@ import CELicensingService from './licensing'
 import { LogsJanitor } from './logs/janitor'
 import { LogsService } from './logs/service'
 import { MediaServiceProvider } from './media'
-import { EventEngine } from './middleware/event-engine'
 import { CEMonitoringService, MonitoringService } from './monitoring'
 import { NLUService } from './nlu/nlu-service'
 import { NotificationsService } from './notification/service'
